Alterações na estrutura dos dados

+++ Banco de Dados/DataBaseGoAesthetic/tabelas/tbAlimentos.sql
@@ -2,6 +2,6 @@
 (
 	[ALM_Id] INT NOT NULL PRIMARY KEY IDENTITY, 
     [IFA_Id] INT NOT NULL, 
-    [ALM_Quantidade] DECIMAL(5, 2) NULL,
+    [ALM_Quantidade] FLOAT NULL,
     CONSTRAINT [FK_tbAlimentos_tbInformacoesAlimentos] FOREIGN KEY ([IFA_Id]) REFERENCES [tbInformacoesAlimentos]([IFA_Id])
 )
